chore: diagnostic fixes — ofacCache scope, eslint plugins, panel prune, ofac rename
Five fixes from the v2.13.0 live diagnostic sweep:

1. fix(sidecar): scope ofacCache via context to prevent ReferenceError
   /api/maritime/vessels and /api/sanctions/* referenced a top-level
   `ofacCache` from inside dispatch(), which had no access to it. The
   bare reference threw, crashing the Node process and taking the whole
   sidecar with it. Pass the cache instance through `context` like the
   rest of the route dependencies, with optional-chain guards so a
   missing cache returns 503 instead of throwing.

2. fix(eslint): declare plugins in test + legacy config blocks
   eslint.config.mjs Block 4 (test files) and Block 5 (legacy large
   files) reference sonarjs/* and unicorn/* rules but never declared
   the plugins, producing 30+ "Definition for rule X was not found"
   errors per lint run. Adding the plugins block matches the working
   pattern from Blocks 2 + 3.

3. chore(config): prune 11 dangling panels from panels.ts
   bitcoin-abuse, cve-tracker, hibp-breaches, ipinfo-lookup,
   mediastack-news, openaq-monitor, phishstats-feed, pulsedive-intel,
   reddit-osint, urlscan-threats, vulners-cve — registered with no
   component file and no rendering logic. Removing both the
   declarations and the panelKeys group references.

4. chore(rename): src/services/sanctions/ofac-search.ts -> ofac-service.ts
   Aligns filename with the v2.13.0 naming convention (ofac-cache.mjs +
   ofac-parser.ts + ofac-service.ts + ofac-types.ts).

Not changed:
- aircraft3d layer was flagged as orphaned but is actually wired via
  DeckGLMap.createAdsb3DLayer() (SimpleMeshLayer 3D variant of ADS-B
  rendered when the toggle is on at zoom ≥ 5). False positive in the
  audit script.

Verified: typecheck zero errors, 25 sanctions tests pass, 131 maritime
tests pass, 97 notification tests pass.
