+++ src-tauri/sidecar/local-api-server.mjs
@@ -4480,9 +4480,9 @@ async function dispatch(requestUrl, req, routes, context) {
     // return the vessels unflagged rather than blocking on a fresh
     // 28MB download mid-request.
     let sanctionedCount = 0;
-    if (ofacCache.indexes) {
+    if (context.ofacCache?.indexes) {
       for (const v of vessels) {
-        const m = ofacCache.matchVessel({ name: v.name, imo: v.imo ?? null, callSign: v.callSign ?? null });
+        const m = context.ofacCache.matchVessel({ name: v.name, imo: v.imo ?? null, callSign: v.callSign ?? null });
         if (m.matched) {
           v.sanctioned = true;
           v.sanctionedReason = m.reason;
@@ -4490,9 +4490,9 @@ async function dispatch(requestUrl, req, routes, context) {
           sanctionedCount++;
         }
       }
-    } else {
+    } else if (context.ofacCache) {
       // Kick off a background refresh so the next call can flag.
-      ofacCache.ensureLoaded().catch(() => {});
+      context.ofacCache.ensureLoaded().catch(() => {});
     }
     const summary = summarizeVesselsSidecar(vessels);
     return json({
@@ -4510,14 +4510,15 @@ async function dispatch(requestUrl, req, routes, context) {
     const q = requestUrl.searchParams.get('q') ?? '';
     const type = requestUrl.searchParams.get('type');
     const limit = Number.parseInt(requestUrl.searchParams.get('limit') ?? '50', 10);
-    if (!q.trim()) return json({ hits: [], meta: ofacCache.getCacheMeta() });
+    if (!context.ofacCache) return json({ hits: [], error: 'sanctions cache unavailable' }, 503);
+    if (!q.trim()) return json({ hits: [], meta: context.ofacCache.getCacheMeta() });
     try {
-      await ofacCache.ensureLoaded();
-      const hits = ofacCache.searchSanctions(q, {
+      await context.ofacCache.ensureLoaded();
+      const hits = context.ofacCache.searchSanctions(q, {
         type: type && ['individual','vessel','aircraft','entity','unknown'].includes(type) ? type : undefined,
         limit: Number.isFinite(limit) ? limit : 50,
       });
-      return json({ hits, meta: ofacCache.getCacheMeta() });
+      return json({ hits, meta: context.ofacCache.getCacheMeta() });
     } catch (error) {
       return json({ hits: [], error: `sanctions-search error: ${error.message ?? error}` }, 502);
     }
@@ -4525,9 +4526,10 @@ async function dispatch(requestUrl, req, routes, context) {
 
   // ── OFAC SDN: all sanctioned vessels (for AIS cross-reference) ──────────
   if (requestUrl.pathname === '/api/sanctions/vessels') {
+    if (!context.ofacCache) return json({ vessels: [], error: 'sanctions cache unavailable' }, 503);
     try {
-      await ofacCache.ensureLoaded();
-      return json({ vessels: ofacCache.getSanctionedVessels(), meta: ofacCache.getCacheMeta() });
+      await context.ofacCache.ensureLoaded();
+      return json({ vessels: context.ofacCache.getSanctionedVessels(), meta: context.ofacCache.getCacheMeta() });
     } catch (error) {
       return json({ vessels: [], error: `sanctions-vessels error: ${error.message ?? error}` }, 502);
     }
@@ -4535,9 +4537,10 @@ async function dispatch(requestUrl, req, routes, context) {
 
   // ── OFAC SDN: all sanctioned aircraft ────────────────────────────────────
   if (requestUrl.pathname === '/api/sanctions/aircraft') {
+    if (!context.ofacCache) return json({ aircraft: [], error: 'sanctions cache unavailable' }, 503);
     try {
-      await ofacCache.ensureLoaded();
-      return json({ aircraft: ofacCache.getSanctionedAircraft(), meta: ofacCache.getCacheMeta() });
+      await context.ofacCache.ensureLoaded();
+      return json({ aircraft: context.ofacCache.getSanctionedAircraft(), meta: context.ofacCache.getCacheMeta() });
     } catch (error) {
       return json({ aircraft: [], error: `sanctions-aircraft error: ${error.message ?? error}` }, 502);
     }
@@ -11269,6 +11272,7 @@ export async function createLocalApiServer(options = {}) {
   loadVerboseState(context.dataDir);
   const routes = await buildRouteTable(context.apiDir);
   const ofacCache = new OfacCache({ dataDir: context.dataDir });
+  context.ofacCache = ofacCache;
 
   const server = createServer(async (req, res) => {
  const requestUrl = new URL(req.url || '/', `http://127.0.0.1:${context.port}`);
