+++ src/services/sanctions/__tests__/ofac.test.mts
@@ -10,7 +10,7 @@ import {
   matchVesselToSanction,
   normalizeVesselName,
   normalizeImo,
-} from '../ofac-search.ts';
+} from '../ofac-service.ts';
 
 const FIXTURE_XML = `<?xml version="1.0" encoding="UTF-8"?>
 <sdnList xmlns="http://tempuri.org/sdnList.xsd">
